+++ hooks/post_install.lua
@@ -5,7 +5,7 @@ local versions = require("lib/versions")
 --- Performs Visual Studio Build Tools installation after mise creates the install directory.
 --- @param ctx {rootPath: string, runtimeVersion: string, sdkInfo: table} Context
 function PLUGIN:PostInstall(ctx)
-    local sdkInfo = ctx.sdkInfo["vsbuild"]
+    local sdkInfo = ctx.sdkInfo["vsbuildtools"]
     local version = sdkInfo.version
     local sdkPath = sdkInfo.path
     local release = versions.resolve(version)